requirements.txt
fastapi
uvicorn[standard]
rawpy
python-multipart
Pillow